+++ apps/client/src/layers/shared/ui/index.ts
@@ -66,6 +66,7 @@ export {
   DropdownMenuRadioGroup,
   DropdownMenuRadioItem,
 } from './dropdown-menu';
+export { HoverBorderGradient } from './hover-border-gradient';
 export { HoverCard, HoverCardTrigger, HoverCardContent } from './hover-card';
 export { Kbd } from './kbd';
 export { Label } from './label';